Move variable out of the script and reuse separately for other scripts

+++ Minecraft server management scripts/RestartMinecraftServer.ps1
@@ -1,10 +1,5 @@
-$minecraftJavaFolderPath = "D:\Minecraft_server"
-$minecraftBedrockFolderPath = "D:\Minecraft_bedrock_server"
-$minecraftSaveArchiveFolder = "D:\minecraft_backup"
 
-$minecraftWorldSubfolderName = "world"
-$minecraftBedrockWorldSubfolderName = "worlds"
-$defaultDate = "12/07/2019"
+. .\Variables.ps1
 
 #internal variables
 $java = Get-Process java -ErrorAction SilentlyContinue
@@ -14,48 +9,59 @@ $hasMineCraftJavaWork = $false
 $hasMineCraftBedrockwork = $false
 
 function MinecraftJavaWorkLoad{
-  if ($java) {
-    Write-Output "kill java window first"
-    Write-Output "Close windows sucessful:" $java.CloseMainWindow() 
-    
-    Start-Sleep 10
-    if (!$java.HasExited) {
-      $java | Stop-Process -Force
+  #kill all the java and javaw, there might be a better way of doing this
+  if($java -Or $javaw){
+    if ($java) {
+      Write-Output "kill java window first"
+      Write-Output "Close windows sucessful:" $java.CloseMainWindow() 
+      
+      Start-Sleep 10
+      if (!$java.HasExited) {
+        $java | Stop-Process -Force
+      }
+      Write-Output "Force close Java completed"
     }
-    Write-Output "Force close Java completed"
+    if ($javaw) {
+      Write-Output "closing javaw window first"
+      Write-Output "Close windows sucessful:" $javaw.CloseMainWindow()
+      
+      Start-Sleep 10
+      if (!$javaw.HasExited) {
+        $javaw | Stop-Process -Force
+      }  
+      Write-Output "Force close Javaw completed"
+    }
+    #Run the Update script
+    if((Test-Path -Path (Join-Path -Path $PSScriptRoot -ChildPath "MinecraftServerUpdate.ps1")-PathType Leaf)){
+      powershell.exe -executionpolicy bypass -file "MinecraftServerUpdate.ps1"
+    }
+    #Set the local variable
+    $Script:hasMineCraftJavaWork = $true
   }
-  if ($javaw) {
-    Write-Output "closing javaw window first"
-    Write-Output "Close windows sucessful:" $javaw.CloseMainWindow()
-    
-    Start-Sleep 10
-    if (!$javaw.HasExited) {
-      $javaw | Stop-Process -Force
-    }  
-    Write-Output "Force close Javaw completed"
-  }    
-  if((Test-Path -Path (Join-Path -Path $PSScriptRoot -ChildPath "MinecraftServerUpdate.ps1")-PathType Leaf)){
-    powershell.exe -executionpolicy bypass -file "MinecraftServerUpdate.ps1" $minecraftJavaFolderPath
-  }
-  $Script:hasMineCraftJavaWork = $true
 }
 
 function MinecraftBedrockWorkload{
+  # kill the bedrock excutable server
   if ($minecraftBedrock) {
     Write-Output "kill bedrock server window first"
     Write-Output "Close windows sucessful:" $minecraftBedrock.CloseMainWindow()
     
-    Start-Sleep 10
+    Start-Sleep 60
     if (!$minecraftBedrock.HasExited) {
       $minecraftBedrock | Stop-Process -Force   
     }
     Write-Output "Force close bedrock server completed"
+    
+    #Run the Update script
+    if((Test-Path -Path (Join-Path -Path $PSScriptRoot -ChildPath "MinecraftBedrockServerUpdate.ps1")-PathType Leaf)){
+      powershell.exe -executionpolicy bypass -file "MinecraftBedrockServerUpdate.ps1"
+    }
+    #Set the local variable
+    $Script:hasMineCraftBedrockwork = $true
   }
-  if((Test-Path -Path (Join-Path -Path $PSScriptRoot -ChildPath "MinecraftBedrockServerUpdate.ps1")-PathType Leaf)){
-    powershell.exe -executionpolicy bypass -file "MinecraftBedrockServerUpdate.ps1" $minecraftBedrockFolderPath
-  }
-  $Script:hasMineCraftBedrockwork = $true
 }
+
+#Start the task
 if(![string]::IsNullOrEmpty($minecraftJavaFolderPath) -and (Test-Path -Path $minecraftJavaFolderPath -PathType Container)){
   MinecraftJavaWorkLoad
 }
@@ -63,23 +69,25 @@ if(![string]::IsNullOrEmpty($minecraftBedrockFolderPath) -and (Test-Path -Path $
   MinecraftBedrockWorkload
 }
 
+#backup the world files
 if(![string]::IsNullOrEmpty($minecraftSaveArchiveFolder) -and (Test-Path -Path $minecraftSaveArchiveFolder -PathType Container)){
-  #backup every 2 weeks if schedule task is running weekly
+  #default backup every 2 weeks if schedule task is running weekly, supporting weekly for now
   $date = [DateTime]::ParseExact($defaultDate, "dd/MM/yyyy", $null)
   if($date -eq $null){
     $date = (Get-Date).Add(15);
   }
+  #change the / will change the frequency of the backup
   if((([math]::Abs([math]::Round(((Get-Date) - $date).day) / 14))%2) -eq 0){ 
       $backupScriptVariables = @()
       $minecraftJavaWorldPath = (Join-Path -Path $minecraftJavaFolderPath -ChildPath $minecraftWorldSubfolderName) 
       $minecraftBedrockWorldPath = (Join-Path -Path $minecraftBedrockFolderPath -ChildPath $minecraftBedrockWorldSubfolderName) 
 
-      if($hasMineCraftJavaWork -and (Test-Path -Path $minecraftJavaWorldPath)){        
+      if($hasMineCraftJavaWork -and (Test-Path -Path $minecraftJavaWorldPath)){
         $backupScriptVariables += $minecraftJavaWorldPath
       }else{
         $backupScriptVariables += ""
       }
-      if($hasMineCraftJavaWork -and (Test-Path -Path $minecraftBedrockWorldPath)){        
+      if($hasMineCraftBedrockwork -and (Test-Path -Path $minecraftBedrockWorldPath)){
         $backupScriptVariables += $minecraftBedrockWorldPath
       }else{
         $backupScriptVariables += ""
@@ -91,14 +99,16 @@ if(![string]::IsNullOrEmpty($minecraftSaveArchiveFolder) -and (Test-Path -Path $
   }
 }
 
-if($hasMineCraftBedrockwork){
-  Invoke-Item "$minecraftBedrockFolderPath\bedrock_server.exe"
-}
+#start servers
+ if($hasMineCraftBedrockwork){
+   Invoke-Item "$minecraftBedrockFolderPath\bedrock_server.exe"
+ }
 
 if($hasMineCraftJavaWork -and (Test-Path -Path "$minecraftJavaFolderPath\StartMinecraftServer.ps1" -PathType Leaf)){
   Write-Output "Starting Minecraft java server via $minecraftJavaFolderPath\StartMinecraftServer.ps1"
   Set-Location $minecraftJavaFolderPath
   powershell.exe -executionpolicy bypass -file StartMinecraftServer.ps1
 }
+
 
 #https://github.com/feletoch/Powershell-Scripts/
